Day 7/requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
pypdf
faiss-cpu